Moving tab layout into shared_seg folder
